Create and upgrade scripts
Upgrade some scripts, like Projectille and BasicSpell, because the prefab Spell wasn't working the way i want, and create some scripts

+++ Player/PlayerHealth.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 public class PlayerHealth : MonoBehaviour
 {
     
-    public int maxHealth = 100; // Set the max health
-    public int currentHealth;
+    public float maxHealth = 100; // Set the max health
+    public float currentHealth;
     public HealthBar healthBar; // calls the HealthBar Script
     
     void Start(){
@@ -14,7 +14,7 @@ public class PlayerHealth : MonoBehaviour
         healthBar.SetMaxHealth(maxHealth); // call the function SetMaxHealth of the HealthBar Script and set the int maxHealth
     }
 
-    public void TakeDamage(int damage){
+    public void TakeDamage(float damage){
         currentHealth -= damage;
         healthBar.SetHealth(currentHealth);
     }
